refactor: modularize UI windows and tweak layout
